Outlining support for target combined directives.
Codegen for a combined construct such as 'target parallel for' requires
outlining to create a target function and a parallel function.  In such
cases we may want to skip adding thread variables (global_tid, bound_tid),
or capture variables only for the target function. This patch modifies
DeclOpenMP to support a CaptureLevel, with the outlining codegen able
to skip unwanted captures at any outlining level.