Modify the 'if' and 'num_threads' clauses to allow capture of expressions within an outlined function.  This is required for combined constructs like 'target parallel', so that the 'target' outlined function is able to capture these expressions.
FIXME: The patch causes the following regression tests to fail.  This seems
to be related to template instantiation code.  The issue is also present
for templated code with a 'parallel for schedule(static, chunk_expr)' so I'm
waiting for Alexey to fix this upstream.
Disabling the failing regression tests for now.