Add support for num_threads and thread_limit clause parsing for combined constructs: target parallel, target parallel for, target teams distribute parallel for.
Rules for num_teams:

If the target directive is combined with a teams directive:
  Return the value in the num_teams clause, if any.
  Otherwise, return 0 to denote the runtime default.

If the target directive is combined with a parallel directive but not a
teams directive, start one team.

Rules for thread_limit:

If the target directive is combined with a teams directive:
  Return the value in the thread_limit clause, if any.

If the target directive is combined with a parallel directive:
  Return the value in the num_threads clause, if any.

If both clauses are set, select the minimum of the two.

If neither teams or parallel combined directives set the number of threads
in a team, return 0 to denote the runtime default.