Specialize NVPTX target codegen based on programming model.  The generic version supports the Fork-Join model of OpenMP.  The SPMD codegen version generates optimized GPU code.  SPMD codegen is used for constructs target parallel, target parallel for, and target teams distribute parallel for.
In SPMD mode parallel threads are activated on entry to the target region
so there can never be a serial region in the kernel.  The amount of parallelism
(i.e., number of threads in the parallel region) remains constant through
the entire target region. We can thus get rid of the warp-pool loop that is
required for the generic codegen.  We also use special GPU runtime calls
to indicate parallel execution at entry.