Rework combined target directive codegen support for host and device.
Use a common codegen for both host and device.
  Each directive has a host call (e.g. EmitOMPTargetDirective) and a device call
  (static function e.g. EmitOMPTargetDeviceFunction) that both call a common function
  (e.g. TargetCodegen) to get the codegen lambda.

Both the host and device codegen use the same lambda.  When we do SPMD codegen, we
now have to specialize in CGOpenMPRuntimeNVPTX.  We introduce a call IsSPMDExecutionMode()
in CGOpenMPRuntimeNVPTX to detect if we are in SPMD execution mode.  We specialize based
on this call, the following functions: registerParallelContext, emitParallelCall,
emitParallelOrTeamsOutlinedFunction, emitProcBindClause, and emitNumThreadsClause.

Finally, for serial execution of the parallel region in a target-parallel or other such
combined directive, we kill all but one thread on the device.