Modified teams codegen to create an outlined function and a simple function call to the outlined function. This happens for all SPMD constructs, like ttdpf, tdpf, ttdpfs, tdpfs. Using the outlining mechanism we can do codegen exclusively in CGStmtOpenMP.cpp.
We would have liked to use the same mechanism of outlining for other
teams constructs such as 'teams', and 'target teams.' However,
outlining of the teams construct creates problems for properly
sharing variables from the team master to the workers that we
were not able to solve. Specifically, for the example below, the
data sharing infrastructure is not able to share the pointer to
'ch' correctly from the team master to all workers.

   int ch = 6;
   #pragma omp target
   #pragma omp teams
   #pragma omp distribute parallel for schedule(static,ch)
   for (int i = 0; i < 1024; i++) {
   }

The data sharing problem does not occur for SPMD constructs.