This is the compiler patch from Tian for implementing Teams reduction on the GPU. I have added support for reduction on several constructs.
Extending the intra-team reduction, this patch creates a function
'kmp_CopyToScratchpadFctPtr scratchFct' that a team master uses to store its single value
to a scratchpad location (a unique element in the scratchpad array indexed by the team id)
in global memory. Additionally, if it is the last team, it uses the function
'kmp_LoadReduceFctPtr ldFct' to load the scratchpad values and reduce it to a single value.
Currently the scratchpad is defined statically in the runtime. A future patch will have to
accept the scratchpad memory location from lomp/libomp on the CPU. If a multi variable
reduction is being performed, the scratchpad is organized with elements for the first
reduction variable followed by the second. There is padding for a 256 byte cache line at
the end of each reduction variable's storage in the scratchpad.

Tian has also reorganized the previously posted intra-team reduction code to pull out the
common parts.

If the construct is a combined construct, all threads in a team participate in the
scratchpad reduction. If, instead, it is a split construct such as #omp teams, then only
the team master participates in the scratchpad reduction since the worker threads are
waiting elsewhere (worker loop) in our scheme.

Calling reduction codegen from the directives is slightly modified. I now pass a
ReductionKind parameter which signifies a 'simd', 'parallel', or 'teams' reduction to
perform. In some cases I pass 'parallel_for_simd' to do both simd and parallel reduction.
In one case I pass 'distribute_parallel_for' to tell the reduction codegen on nvptx to
do the parallel reduction but not update the final value. This is because a subsequent
teams reduction call will perform this action. Unfortunately I had to resort to this
convoluted method because the 'distribute' part does not generate an outlined function.