This patch enables data sharing support in SPMD constructs such as 'target parallel' and 'target teams distribute parallel for'.
    It sets up the ParallelNestingLevel correctly when generating the
      SPMD construct.
    It sets up the ParallelNestingLevel correctly when generating the
    coalesced schedule for the SPMD construct 'target teams distribute
      parallel for'.
    It codegens calls to nested parallel directives correctly.

To simplify the code, I've replaced 'emitParallelOrTeamsOutlinedFunction' with two
functions 'emitTeamsOutlinedFunction' and 'emitParallelOutlinedFunction'.

When setting up an SPMD construct, I inspect the code to see if data sharing may
occur in the target region (nested parallel/simd or calls to an external function).
If so, I set RequiresDataSharing to true in the call __kmpc_spmd_kernel_init().